Refactor: Extract SSDPDiscoveryService from SonosController
Phase 1 (Step 2): Extract SSDP discovery logic into dedicated service.

- Create Infrastructure/SSDPDiscoveryService.swift (180 lines)
- Extract discovery protocol, response parsing, and device info fetching
- Introduce SSDPDiscoveryResult as clean data transfer object
- Refactor SonosController to use discovery service
- Remove parseSSDPResponse() and fetchDeviceInfo() from SonosController
- Reduce SonosController from 1,654 to 1,577 lines (-77)

Benefits:
- SSDP discovery logic is now independently testable
- Clear separation between discovery and device management
- Better encapsulation with actor isolation for thread safety
- Constants properly organized in nested enum

Related to P1 Architecture task in ROADMAP.md:61
